+++ src/shared/ui/SearchBar/hooks/useSearchQuery.tsx
@@ -1,0 +1,1 @@
+const useSearchQuery = () => {};
